commit per passare su un altro pc

+++ digitalterritory/src/main/java/it/unicam/cs/ids/digitalterritory/dto/OsmResponse.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 @Data
 public class OsmResponse {
     @JsonProperty("place_id")
-    private int placeId;
+    private long placeId;
     private String licence;
     @JsonProperty("osm_type")
     private String osmType;
